refactor(types): centralize CommandType and fix SequenceTask.action typing
