+++ src/types/index.ts
@@ -1,5 +1,20 @@
 export type RobotConnectionStatus = "connected" | "disconnected" | "connecting";
 
+// Direct control commands — matches backend routes/commands.py CommandType enum exactly
+export type CommandType =
+  | "move_forward"
+  | "move_backward"
+  | "move_left"
+  | "move_right"
+  | "move_up"
+  | "move_down"
+  | "rotate_left"
+  | "rotate_right"
+  | "stop"
+  | "go_home"
+  | "gripper_open"
+  | "gripper_close";
+
 export type RobotOperationalStatus = "active" | "idle" | "paused" | "stopped" | "error";
 
 export type ScheduledTask = "home" | "retract" | "custom";
